feat: DialogComponentsに新しいZapアイテムのHTML生成機能を追加し、ZapListUIでの利用を最適化
